refactor: separate workshop access and revision sessions
