+++ ui/src/pages/skill-workshop/view-types.ts
@@ -5,17 +5,12 @@ import type {
   SkillWorkshopProposal,
   SkillWorkshopStatusFilter,
 } from "../../lib/skill-workshop/index.ts";
+import type { SkillWorkshopAccess } from "./access.ts";
 import type { SkillWorkshopSelfLearning } from "./self-learning.ts";
 import type { SkillWorkshopHistoryScanState } from "./state.ts";
 
 export type SkillWorkshopProps = {
-  access: {
-    canEvaluate: boolean;
-    canApply: boolean;
-    canRevise: boolean;
-    canReject: boolean;
-    canScanHistory: boolean;
-  };
+  access: SkillWorkshopAccess;
   loading: boolean;
   error: string | null;
   inspectingKey: string | null;
